Fix warning.
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@20859 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/tapi_rpc/tapi_rpc_internal.h
@@ -184,7 +184,7 @@
  * @param _var      variable with return value
  */
 #define CHECK_RETVAL_VAR_IS_ZERO(_func, _var) \
-    CHECK_RETVAL_VAR(_func, _var, ((_var) != 0), -1)
+    CHECK_RETVAL_VAR(_func, _var, ((_var) != 0), (uint32_t)-1)
 
 /** Return with check (for functions returning zero value) */
 #define RETVAL_ZERO_INT(_func, _retval) \
